% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_quiz.R
\name{build_results_table}
\alias{build_results_table}
\title{Title}
\usage{
build_results_table(state, county, ans_20, ans_50, ans_100, ans_1000)
}
\arguments{
\item{sel_state}{}
}
\value{

}
\description{
Title
}
