% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{close_connections}
\alias{close_connections}
\title{close_connections
Close any hung DBI connections}
\usage{
close_connections()
}
\value{

}
\description{
close_connections
Close any hung DBI connections
}
