% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_feedback.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Title}
\usage{
connect_to_db(
  drv = RMySQL::MySQL(),
  username = Sys.getenv("MYSQL_USERNAME"),
  password = Sys.getenv("MYSQL_PASSWORD"),
  host = Sys.getenv("MYSQL_HOST"),
  port = 3306,
  dbname,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
