% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{get_data}
\alias{get_data}
\title{Get data needed for app to function}
\usage{
get_data()
}
\description{
While this function has no arguments, there are
environment variables that are import to its functioning
}
\details{
C19R_CASES_DIR Alternative path to NYT COVID19 case data
C19R_RISK_DATA Alternative path to pre-computed risk data
C19R_EXTERNAL_UPDATES Prevent app from update NYT case data
}
