% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map.R
\name{maplabs}
\alias{maplabs}
\title{maplabs}
\usage{
maplabs(riskData)
}
\arguments{
\item{riskData}{Risk data df given to [leaflet::addPolygons()]}
}
\value{
[htmltools::HTML] vector of map labels to user
}
\description{
maplabs
}
