% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map.R, R/utils_helpers.R
\name{roundUpNice}
\alias{roundUpNice}
\title{Title}
\usage{
roundUpNice(x, nice = c(1, 2, 4, 5, 6, 8, 10))

roundUpNice(x, nice = c(1, 2, 4, 5, 6, 8, 10))
}
\arguments{
\item{nice}{}
}
\value{



}
\description{
Title

Title
}
