% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_quiz.R
\name{save_quiz_results}
\alias{save_quiz_results}
\title{save_quiz_results}
\usage{
save_quiz_results(
  db,
  geoid,
  data_ts,
  p20,
  p50,
  p100,
  p1000,
  g20,
  g50,
  g100,
  g1000,
  ip,
  lat,
  long,
  utm_source,
  utm_medium,
  utm_content,
  utm_campaign
)
}
\arguments{
\item{db}{Database, from global env (see [loadDataOnStart()])}

\item{geoid}{Geoid of location selected}

\item{data_ts}{Timestamp of quiz submission}

\item{p20}{Predicted event size = 20}

\item{p50}{Predicted event size = 50}

\item{p100}{Predicted event size = 100}

\item{p1000}{Predicted event size = 1000}

\item{g20}{Guess event size = 20}

\item{g50}{Guess event size = 50}

\item{g100}{Guess event size = 100}

\item{g1000}{Guess event size = 1000}

\item{ip}{User ip if known, from `globals$ip()`}

\item{lat}{User latitude, 0 if unknown.  Typically from `globals$latitude()`}

\item{long}{User longitude, 0 if unknown.  Typically from `globals$longitude()`}

\item{utm_source}{`utm_source=` value from query string or 'NULL'}

\item{utm_medium}{`utm_medium=` value from query string or 'NULL'}

\item{utm_content}{`utm_content=` value from query string or 'NULL'}

\item{utm_campaign}{`utm_campaign=` value from query string or 'NULL'}
}
\value{

}
\description{
save_quiz_results
}
