% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{save_willingness}
\alias{save_willingness}
\title{Title}
\usage{
save_willingness(
  source,
  asc_bias,
  event_size,
  answer,
  ip,
  vacc_imm,
  latitude,
  longitude,
  utm_source = "NULL",
  utm_medium = "NULL",
  utm_content = "NULL",
  utm_campaign = "NULL"
)
}
\arguments{
\item{source}{}

\item{asc_bias}{}

\item{event_size}{}

\item{answer}{}

\item{ip}{}

\item{vacc_imm}{}

\item{latitude}{}

\item{longitude}{}

\item{utm_source}{}

\item{utm_medium}{}

\item{utm_content}{}

\item{utm_campaign}{}
}
\value{

}
\description{
Title
}
