% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_inplace.R
\name{setCircleMarkerRadius}
\alias{setCircleMarkerRadius}
\title{Title}
\usage{
setCircleMarkerRadius(map, layerId, radius, data = leaflet::getMapData(map))
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Title
}
