% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_inplace.R
\name{setCircleMarkerStyle}
\alias{setCircleMarkerStyle}
\title{setCircleMarkerStyle
from https://github.com/rstudio/leaflet/pull/598}
\usage{
setCircleMarkerStyle(
  map,
  layerId,
  radius = NULL,
  stroke = NULL,
  color = NULL,
  weight = NULL,
  opacity = NULL,
  fill = NULL,
  fillColor = NULL,
  fillOpacity = NULL,
  dashArray = NULL,
  options = NULL,
  data = leaflet::getMapData(map)
)
}
\arguments{
\item{map}{Leaflet map object}

\item{layerId}{Layer id}

\item{radius}{Cicle radius}

\item{stroke}{Stoke}

\item{color}{Stroke color}

\item{weight}{Stroke weight (width)}

\item{opacity}{Stroke opacity}

\item{fill}{Circle fill}

\item{fillColor}{Circle fill color}

\item{fillOpacity}{Circle fill opacity}

\item{dashArray}{NULL}

\item{options}{Leaflet options}

\item{data}{Leaflet map data}
}
\value{

}
\description{
setCircleMarkerStyle
from https://github.com/rstudio/leaflet/pull/598
}
