% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_inplace.R
\name{setShapeStyle}
\alias{setShapeStyle}
\title{setShapeStyle
from https://github.com/rstudio/leaflet/pull/598}
\usage{
setShapeStyle(
  map,
  data = leaflet::getMapData(map),
  layerId,
  stroke = NULL,
  color = NULL,
  weight = NULL,
  opacity = NULL,
  fill = NULL,
  fillColor = NULL,
  fillOpacity = NULL,
  dashArray = NULL,
  smoothFactor = NULL,
  noClip = NULL,
  options = NULL
)
}
\arguments{
\item{options}{}
}
\value{

}
\description{
setShapeStyle
from https://github.com/rstudio/leaflet/pull/598
}
