% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map.R
\name{sidebar_text}
\alias{sidebar_text}
\title{Generate sidebar text}
\usage{
sidebar_text(geo_specific = "a county")
}
\arguments{
\item{geo_specific}{String for geospecific region description}
}
\value{
HTML for sidebar
}
\description{
Generate sidebar text
}
