% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_quiz.R
\name{build_results_table}
\alias{build_results_table}
\title{build_results_table}
\usage{
build_results_table(state, county, ans_20, ans_50, ans_100, ans_1000)
}
\arguments{
\item{state}{State selection from dropdown}

\item{county}{County selection from dropdown}

\item{ans_20}{User answer for the event size = 20 slider}

\item{ans_50}{User answer for the event size = 50 slider}

\item{ans_100}{User answer for the event size = 100 slider}

\item{ans_1000}{User answer for the event size = 1000 slider}
}
\description{
build_results_table
}
