% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map.R, R/utils_helpers.R
\name{calc_risk}
\alias{calc_risk}
\title{Title}
\usage{
calc_risk(I, n, USpop, scaling_factor = 10/14)

calc_risk(I, n, USpop, scaling_factor = 10/14)
}
\arguments{
\item{scaling_factor}{}
}
\value{



}
\description{
Title

calc_risk
}
