% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_feedback.R
\name{connect_to_db}
\alias{connect_to_db}
\title{connect_to_db}
\usage{
connect_to_db(
  drv = RMySQL::MySQL(),
  username = Sys.getenv("MYSQL_USERNAME"),
  password = Sys.getenv("MYSQL_PASSWORD"),
  host = Sys.getenv("MYSQL_HOST"),
  port = 3306,
  dbname,
  ...
)
}
\arguments{
\item{drv}{DBI-compatible database driver}

\item{username}{login username}

\item{password}{login password}

\item{host}{database hostname or IP}

\item{port}{database port}

\item{dbname}{name of database/table to use}

\item{...}{}
}
\value{
`pool::dbPool` obj`
}
\description{
Connect to database and return pool object
}
\examples{
connect_to_db(drv = RMySQL::MySQL(), username = "root", password = "root")
}
