% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_feedback.R, R/utils_helpers.R
\name{make_resp_slider}
\alias{make_resp_slider}
\title{Title}
\usage{
make_resp_slider(input, label)

make_resp_slider(input, label)
}
\arguments{
\item{label}{}
}
\value{



}
\description{
Title

Title
}
