% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_feedback.R
\name{make_resp_slider}
\alias{make_resp_slider}
\title{make_resp_slider}
\usage{
make_resp_slider(input, label)
}
\arguments{
\item{input}{`{shiny}` input id}

\item{label}{label for input}
}
\value{
Slider input
}
\description{
Create slider inputs for Risk Quik, range from 0-100%
}
