% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map.R
\name{roundUpNice}
\alias{roundUpNice}
\title{roundUpNice}
\usage{
roundUpNice(x, nice = c(1, 2, 4, 5, 6, 8, 10))
}
\arguments{
\item{x}{Number to round}

\item{nice}{log10 ingrements to try to round to nearest}
}
\value{
Rounded number
}
\description{
roundUpNice
}
\examples{
roundUpNice(124207)
}
