% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_feedback.R
\name{save_willingness}
\alias{save_willingness}
\title{save_willingness}
\usage{
save_willingness(
  db,
  source,
  asc_bias,
  event_size,
  answer,
  ip,
  vacc_imm,
  latitude,
  longitude,
  utm_source = "NULL",
  utm_medium = "NULL",
  utm_content = "NULL",
  utm_campaign = "NULL"
)
}
\arguments{
\item{db}{Database or pool object}

\item{source}{Where is willingness being be asked, e.g. "game" or "map"}

\item{asc_bias}{Ascertainment bias when asked on map, set to -1 for other uses}

\item{event_size}{Event size when asked on map, set to -1 for other uses}

\item{answer}{Willingness answer, should come from the willingness scale}

\item{ip}{User ip if known, typically from `globals$ip()``}

\item{vacc_imm}{Vaccine immunity overlay setting if from map}

\item{latitude}{User latitude, 0 if unknown.  Typically from `globals$latitude()`}

\item{longitude}{User longitude, 0 if unknown.  Typically from `globals$longitude()`}

\item{utm_source}{`utm_source=` value from query string or 'NULL'}

\item{utm_medium}{`utm_medium=` value from query string or 'NULL'}

\item{utm_content}{`utm_content=` value from query string or 'NULL'}

\item{utm_campaign}{`utm_campaign=` value from query string or 'NULL'}
}
\description{
Save user willingness input from prompts on maps and quiz tab
}
