% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_inplace.R
\name{setCircleMarkerStyle}
\alias{setCircleMarkerStyle}
\title{Title}
\usage{
setCircleMarkerStyle(
  map,
  layerId,
  radius = NULL,
  stroke = NULL,
  color = NULL,
  weight = NULL,
  opacity = NULL,
  fill = NULL,
  fillColor = NULL,
  fillOpacity = NULL,
  dashArray = NULL,
  options = NULL,
  data = leaflet::getMapData(map)
)
}
\arguments{
\item{map}{}

\item{layerId}{}

\item{radius}{}

\item{stroke}{}

\item{color}{}

\item{weight}{}

\item{opacity}{}

\item{fill}{}

\item{fillColor}{}

\item{fillOpacity}{}

\item{dashArray}{}

\item{options}{}

\item{data}{}
}
\value{

}
\description{
Title
}
