% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_inplace.R
\name{setShapeLabel}
\alias{setShapeLabel}
\title{setShapeLabel
from https://github.com/rstudio/leaflet/pull/598}
\usage{
setShapeLabel(
  map,
  data = leaflet::getMapData(map),
  layerId,
  label = NULL,
  options = NULL
)
}
\arguments{
\item{options}{}
}
\value{

}
\description{
setShapeLabel
from https://github.com/rstudio/leaflet/pull/598
}
