% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_inplace.R
\name{setShapeLabel}
\alias{setShapeLabel}
\title{Title}
\usage{
setShapeLabel(
  map,
  data = leaflet::getMapData(map),
  layerId,
  label = NULL,
  options = NULL
)
}
\arguments{
\item{map}{}

\item{data}{}

\item{layerId}{}

\item{label}{}

\item{options}{}
}
\value{

}
\description{
Title
}
