% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_inplace.R
\name{setShapeStyle}
\alias{setShapeStyle}
\title{Title}
\usage{
setShapeStyle(
  map,
  data = leaflet::getMapData(map),
  layerId,
  stroke = NULL,
  color = NULL,
  weight = NULL,
  opacity = NULL,
  fill = NULL,
  fillColor = NULL,
  fillOpacity = NULL,
  dashArray = NULL,
  smoothFactor = NULL,
  noClip = NULL,
  options = NULL
)
}
\arguments{
\item{map}{}

\item{data}{}

\item{layerId}{}

\item{stroke}{}

\item{color}{}

\item{weight}{}

\item{opacity}{}

\item{fill}{}

\item{fillColor}{}

\item{fillOpacity}{}

\item{dashArray}{}

\item{smoothFactor}{}

\item{noClip}{}

\item{options}{}
}
\value{

}
\description{
Title
}
