% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_inplace.R
\name{setShapeStyle}
\alias{setShapeStyle}
\title{setShapeStyle
from https://github.com/rstudio/leaflet/pull/598}
\usage{
setShapeStyle(
  map,
  data = leaflet::getMapData(map),
  layerId,
  stroke = NULL,
  color = NULL,
  weight = NULL,
  opacity = NULL,
  fill = NULL,
  fillColor = NULL,
  fillOpacity = NULL,
  dashArray = NULL,
  smoothFactor = NULL,
  noClip = NULL,
  options = NULL
)
}
\arguments{
\item{map}{Leaflet map object}

\item{data}{Leaflet map data}

\item{layerId}{Layer to manipulate}

\item{stroke}{Stroke (bool)}

\item{color}{Shape color}

\item{weight}{Stroke weight (line width)}

\item{opacity}{Stroke opacity}

\item{fill}{Shape fill (bool)}

\item{fillColor}{Fill color}

\item{fillOpacity}{Fill opacity}

\item{dashArray}{NULL}

\item{smoothFactor}{Should shapes and colors be smooth}

\item{noClip}{Allow clipping}

\item{options}{Leaflet options}
}
\value{

}
\description{
setShapeStyle
from https://github.com/rstudio/leaflet/pull/598
}
