% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_feedback.R
\name{str_or_unk}
\alias{str_or_unk}
\title{str_or_unk}
\usage{
str_or_unk(obj)
}
\arguments{
\item{obj}{object to voncert}
}
\value{
converted object
}
\description{
Convert null objects to "Unknown"
}
\examples{
\dontrun{
 str_or_unk(NULL)
 str_or_unk("example")
}
}
